[lldb] Fix tests to resolve symlinks when checking paths
The inferior used in the process launch test prints out its working
directory using the `getcwd` function, which is not allowed to return
symbolic links in the path components. When testing against the output
from `getcwd` we should resolve the full path to match the expected
output.

The source manager test sets a breakpoint on a main-copy.c file that is
copied into the build output directory. When the source manager seems to
resolve this file to its real location. When testing the output from
the source cache we need to resolve the expected path to remove
symlinks.