Moved LLDBServerPlugin into LLDB library.
This patch formalizes the plug-ins and moves the code into the LLDB
library lldbPluginProcessGDBRemote. This allows the sever objects
GDBRemoteCommunicationServerLLGS to own a list of GPU plug-ins.

Removed the process stopped callback and we now iterate over all of the
installed GPU plug-ins in the GDBRemoteCommunicationServerLLGS class
itself. This will allow more functionality to be implemented for GPU
plugins.