Added "bool load_libraries;" to the GPUActions structure.
This patch makes it so that a breakpoint in the native process can cause
shared libraries to be loaded in the GPU process. When GPU plug-ins
respond to the LLDBServerPlugin method:

  GPUPluginBreakpointHitResponse
  LLDBServerPlugin::BreakpointWasHit(GPUPluginBreakpointHitArgs &args);

They can now set the GPUPluginBreakpointHitResponse.load_libraries to
true. The GPU plug-in should already hvae notified LLDB that it is
stopped prior to sending this to ensure that no progress is made on the
GPU while shared libraries are loaded and breakpoints get resolved.

Code was added to Target.h that allows a target to know about all of the
installed GPU plug-ins. This allows a native process to make calls on
the GPU process, and also for the GPU process to get the native target.
This will allow code to resume both targets when one gets resumed if
that is the preferred methodology for the native process and GPU. It
also allows us to stop the native target from the GPU target and vice
versa.

The LLDBServerPluginMockGPU now tests that breakpoints by address work
by setting a brekapoint by address from the "gdb_shlib_load" symbol that
is requested by the "gpu_initialize" breakpoint.