Add tests for amdgpu lldb-server plugin (#7)
* Add tests for amdgpu lldb-server plugin

This commit adds some basic unit tests for the AMDGPU plugin.
It modifies the test configuration to expose the HIPCC compiler
to the make file when the AMDGPU plugin is enabled. It also modifies
the dotest framework so that we can compile .hip files using the
Makefile.rules like we do for other tests.

The unit tests compile a simple hip program and validate that we do
create the gpu target succesfully and that we can hit a breakpoint
in the gpu code.

* Use dynamic line number for breakpoint

* Use LLDB_ENABLE_AMDGPU_PLUGIN configuration variable

* Rename getRocmArgs to getHipccArgs