[LLDB][AMDGPU] Set gpu loader breakpoint by name (#20)
Currently we wet the gpu loader breakpoint by address in the
`amd_dbgapi_insert_breakpoint_callback` callback. This means we need to
halt the CPU in order to correctly set the breakpoint. When we halt the
CPU it shows up as a public stop to the user and interrupts the
debugging flow.

This PR changes it so that we set the loader breakpoint by name when we
first create the gpu connection. We set the breakpoint on the
`rocr::_loader_debug_state` function which was discovered by looking up
the address passed into the callback.

This is a bit of a hack since the function name could potentially change
over time or may be unavailable if the runtime was statically linked.
But we use it for now to make the debugger experience better until we
can find a more permanant solution.