[amd] Add tests for reading and writing gpr registers
This commit adds two tests for registers

  * Verify we can read and write all vgprs (v0-v255) and sgprs (s0-101)
  * Verify that updating a subset of lanes in a vector register works

To check the reading and writing of gprs we initialize them to a known
value using inline assembly. Then we verify that we read the expected
value after hitting the initial gpu breakpoint and that we can write
and updated value and read it back.

The second test (`test_vector_lane_read_write`) makes sure that we
can update the elements of a vector register to different values. It is
needed because the first test is splatting the same value to all lanes
in a wave.

Since we have such a large number of registers I added a script
(`gen.py`) to help with generating the assembly to initialize the values
and the python code to hold the expected values. This is much easier
than writing it all out by hand.

We did not test the agpr registers in this commit because we seem to
have a bug reading the correct values. Will fix the bug and add tests in
a future patch.