[GPU][amd] Fix disassembler for amdgpu plugin (#24)
The AMDGPU disassembler needs to know the precise cpu in order to disassemble correctly. This is because the triple for these targets is not enough to identify the target architecture. The disassembler in lldb was already passing down the target cpu correctly so we just need to hook it up for the amd plugin.

1. Query the target's architecture for the `GetClangTargetCPU()` value and use that for disassembly if it has a valid value.
2. Pass the cpu type and cpu sub-type back from the gdbserver for amdgpu and use it to create the target architecture.

The cpu type and sub-type are currently only enabled for MachO, but we need them for amdgpu elf files as well.

There was an existing target property (`GetDisassemblyCPU`) that is used to control the cpu for the disassembler. We wanted to keep the property code simple to just read the property value. So now when the target's arch spec is updated we will use that to set the target property if there is not an existing value.