[GPU][AMD] Only report new modules to the dynamic loader (#29)
This commit changes the `GetGPUDynamicLoaderLibraryInfos` callback so that we
only return the new modules when `full` is false. I initially thought it would
be a simple change, but after investigating I decided to create a new class to
manage the loaded module state.

The state management is a bit complicated because the gpu debug library only
returns the full set of loaded code objects, but lldb wants to be able to just
get the list of modules that have changed. Additionally, we need to be able to
track when a module has been unloaded. There is no separate event for that so we
have to track it by detecting when a previously loaded module is no longer in
the current list of active code objects returned by the debug library. The
GpuModuleManager class hides those details and provides a convenient interface
that works well to bridge the gap between lldb and the debug library.

As part of this change I enabled unit tests for the lldb-server plugins using
the same cmake variables that we use to enable the plugins. I also fixed a
memory leak where we were previously not freeing the code object uri string.