Call notify on StateSync condition variable when resuming (#39)
We are currently missing a call to notify threads waiting on the
StateSync condition variable when resuming the process. This causes the
native process to miss out on notifications that the gpu process has
resumed when waiting for the gpu to process to hit its requested
stop id when `wait_for_gpu_process_to_resume` is true in GPUActions.