Add new constructor to AddressSpec.
Add the ability to create a AddressSpec class with an address and memory
space integer identifier with optional thread. This allows internal
clients that parse debug info that contains the address space integer
identifier to use the AddressSpec to read memory. Prior to this clients
had to use the address space name.