[GPU][AMD] Increase the size of the max CIE augmentation string
The CIE storage uses a fixed size buffer to parse the augmentation string. The
AMDGPU backend uses an augmentation string that exceeds the current
size, which causes the parser to fail. This PR increases the storage
size to fit the AMDGPU augmentation string and adds a test to validate
that we can parse it successfully.