[GPU][AMD] Refactor initialization code and delay connection (#41)
This PR refactors the amdgpu plugin initialization logic to make it more
flexible so that we can choose when to initialize the debug library separately
from attaching the process and creating the connection.

The change mostly shuffles around and cleans up existing code, but we
now also explicitly track the state of the amd debug library (e.g.
initialized, attached, runtime-initialized) so that we can use that to
guide decisions around when to generate the debug connection.

We also fixed the connection logic after the changes in f1343a4.  It now
delays the connection until after the initial stop that occurs for the
native process when it first ptrace attaches. The initial stop does not
handle the GPUActions so we need to wait for a later stop.

I played with delaying the connection even further to when the rocm
runtime is initialized. That works, but it makes it awkward to use the
debugger to set gpu breakpoints. The runtime is initialized on demand so
there is not always a good place to set a cpu breakpoint where we can
halt the process and create the gpu breakpoints. If we change the
debugger to propagate the breakpoints from the cpu to gpu then this will
not be an issue because we can set the breakpoints before the gpu target
is created.