[NFC][AMD][GPU] Return processed events from process_event_queue
This commit modifies the `process_event_queue` function to return the
set of events that were processed. It also pushes the acknowledgment of
the event into the `handleGPUEvent` function so that all events will be
handled once the call to `process_event_queue` returns.

This is used in an upcoming change to make it easier to implement
lanes-as-threads by having an explicit check to know when we have hit a
wave breakpoint instead of checking for the modification of an internal
data structure.