[AMD][GPU] Return processed events from process_event_queue (#49)
This commit modifies the process_event_queue function to return the set of events that were processed. It also pushes the acknowledgment of the event into the handleGPUEvent function so that all events will be acknowledge once they are handled.

We also modify the process_event_queue function to give the caller the choice of how they want to handle the until_event target event. We can either not handle the event (which was the previous semantics) or choose to handle the event (which is the new default). We were not taking advanatge of delaying the handling of the event so I changed the default to handle the target event.

This is used in an upcoming change to make it easier to implement lanes-as-threads by having an explicit check to know when we have hit a wave breakpoint instead of checking for the modification of an internal data structure.