[GPU] Added identifier for GPU action (#60)
After merging #48 I notice that my test is failing in the upcoming PR: #46.

The issue is that having a map of gpu actions with stop_id as a value will create a situation where unique GPUactions that have the same stop_id and name are getting ignored.
Having a unique identifier for every gpu action will prevent this situation from happening while still ignoring same gpu action to be executed twice.