[GPU][AMD] Cleanup the reading from the notifier fd
The debug library writes some data to the notifier file descriptor to
indicated an event is ready to be processed. This commit cleans up the
code that reads the data from this descriptor.

It contains the following changes

  - Remove the unused double nested loop
  - Only loop on `read` for EINTR errors
  - Stop overriding the unneeded `HandleEventFileDescriptorEvent`
  - Rename the processGPUEvent function to HandleNotifierDataReady
  - Add more logging
  - Log non-fatal errors about forward progress handling

The extra layer of `do {...} while(true)` loop around the read was not
needed because we would unconditionally break out of the loop.

The logic for re-doing the `read` system call on the notifier descriptor
was changed to only loop when we get an EINTR error code. This avoids
looping on EOF and other errors which could lead to infinite busy loops.

The HandleEventFileDescriptorEvent override was unused because we
explicitly register a callback with `RegisterReadObject` on the main
loop. Removing it cleans up the code and makes it obvious it is not
needed.