[GPU] GPUActions are ignored on first stop-reply packet fix (#46)
GPUActions were not handled on the first stop-reply packed that is returned when we launch a process.

Modified the `LaunchProcess` method to capture and return the first stop-reply packet from the `vRun` command instead of discarding it. In `ProcessGDBRemote::DoLaunch`, we now use this captured launch response directly if it's a valid stop-reply packet, so GPU actions in the launch response are processed instead of ignored..