[GPU][AMD] Compute the number of threads in a wave correctly (#63)
Previously we used the exec mask to determine the number of lanes in a
wave. This would incorrectly compute the number of lanes if the wave is
stopped at a divergent place in the code (i.e. stopped in a place where
not all lanes in the wave are active).

We fix this by querying the dispatch info for the wave to compute the
total number of lanes across the entire block. Then we adjust the number
of threads for the last lane in the block to account for the case where
it is not a full wave worth of lanes.

Added a new test to verify that we correctly compute the number of lanes
even when stopped at a divergent breakpoint.