Add support for unloading libraries and loading at file addresses.
This patch adds support for unloading shared libraries and for loading a
library exactly at the addresses already found in the file. This is
similar to how the DynamicLoaderStatic works where it loads each shared
library at the file addresses, or with no changes or slides. This is
done by not specifying a GPUDynamicLoaderLibraryInfo.load_address or a
GPUDynamicLoaderLibraryInfo.loaded_sections. When neither are specified,
then they will get loaded statically.