[lldb] Handle deref of a register location
Handles DW_OP_deref* operations on register location descriptions as
described by the heterogenious debugging
[extensions](https://rocm.docs.amd.com/projects/llvm-project/en/latest/LLVM/llvm/html/AMDGPUDwarfExtensionsForHeterogeneousDebugging.html#a-2-5-4-4-4-register-location-description-operations).

Specifically, for register locations it states

> These operations obtain a register location. To fetch the contents of
> a register, it is necessary to use DW_OP_regval_type, use one of the
> DW_OP_breg* register-based addressing operations, or use DW_OP_deref* on
> a register location description.

In lldb for operations that create register locations (e.g. DW_OP_reg0)
we will read the value of the register and push it on the stack. This
means that when we go to "deref" that location we have actually already
read the contents and just need to reset the location to the default
(Memory) and keep the value on the top of the stack.

My understanding is that this is the intended behavior from dwarf5 as
well and is not a change in behavior.