[GPU] Use AddressSpec in ReadMemory for GetValueAsData
This commit modifies the Value class to use the ReadMemory api that
takes an AddressSpec for the address. This is a preliminary commit for
enabling address spaces inside values.

Most of the work here was on the testing side, which requires having a
proper CompilerType for the value so it can exercise the load.