[wip] Add support for DW_OP_LLVM_form_aspace_address
This commit adds support for handling the llvm dwarf6 extension opcodes
and implements the DW_OP_LLVM_form_aspace_address operation. This
is implemented by adding a `std::optional` address space to the Value
class and setting that value based on the dwarf opcode.

We also need to modify the MemoryRead used by the value when loading the
contents to make use of the address space stored in the value.

We currently do not handle pass the address space to other memory reads
inside the dwarf evaluator. This can be done as a future extension by
piping the address space through the layers, but is not currently needed
for evaluating simple variables.