[GPU][AMD] Add initial support for dwarf6 llvm extensions
This commit adds support for the DW_OP_LLVM_form_aspace_address and
DW_OP_LLVM_offset opcodes in the dwarf evaluator. Support for these
opcodes allow us to view varibles with the `frame variable` command on
the gpu target.

The  DW_OP_LLVM_form_aspace_address requires keeping track of a values
address space and is implemented by adding a `std::optional` address space
to the Value class which is set based on the dwarf opcode.

We also need to modify the MemoryRead used by the value when loading the
contents to make use of the address space stored in the value and also
propagate the address space from the parent ValueObject to the
ValueObjectChild when computing its address.

We currently do not handle passing the address space to other memory reads
inside the dwarf evaluator. This can be done as a future extension by
piping the address space through the layers.

The tests for global variables are currently marked xfail and need to be
fixed in a future change.