[GPU][AMD] Fix lldb hipcc tests to link using libstdc++
This commit changes the test Makefile.rules to force linking with
libstdc++ when building with hipcc.

An upstream [change](https://github.com/llvm/llvm-project/pull/164462)
modified the test link commands to add -nostdlib++ to avoid linking both the
internal llvm c++ library and the external system c++ standard library.
This was causing linking failures when building with hipcc because it
expects to use the system libstdc++ library.